contextual bundle files moved around to match the new package structure in master
